Danbooru: Refactor DanbooruAuthInterceptor
